7/21/25
Attempted to figure out Certification and SSL issues but Firefox wont accept the SelfSignedMitmManager rootCA created on the fly.

Fixed some small bugs that were present on the front end and got that fixed so its displaying HTTP domains that are blocked, the IP address and Protocol used still arent being represented correctly as this cant be obtained due to not being able to get to the datagram.